+++ frontend/src/services/driver.service.js
@@ -4,8 +4,8 @@ const driverService ={
     createDriver(driverData){
         return api.post('/driver/create',driverData)
     },
-    getDrivers(){
-        return api.get('/driver/get-drivers')
+    getDrivers(userId){
+        return api.get('/driver/get-drivers', { params: { userId } })
     },
     getDriver(id){
         return api.get(`/driver/get-driver/${id}`)
@@ -13,8 +13,8 @@ const driverService ={
     updateDriver(id,updatedData){
         return api.put(`/driver/update-driver/${id}`, updatedData)
     },
-    deleteDriver(id){
-        return api.delete(`/driver/delete-driver/${id}`)
+    deleteDriver(id, userId){
+        return api.delete(`/driver/delete-driver/${id}`, { params: { userId } })
     }
 }
